* fix: "공지사항 컨트롤러,Dto 수정"

* fix: "제품 리스트 및 상세 Dto 수정"

* fix: "제품 리스트 및 상세 Dto 2차 수정"

* fix: "제품 리스트 및 상세 Dto 최종 수정"

* Feat: "발주 상품 CRUD 구현"

주문 상품 조회 등 구현 및 오더 엔티티 수정, serialVersionUID 통일

* feat: "공지사항 유저 권한 체크 구현"

* feat: "엑셀/S3/상품 관리자 체크 구현"

* feat: "공지/파일/주문 권한 설정"

* Feat: "로그인 기능 개선"

* Feat: "로그인 기능 개선"

* Feat: "로그인 기능 개선"

* feat: "주문 상세 승인/거절 구현"

* feat: "엑셀 파일 다운방식 변경 / 주문 요청 기능 구현"

* Fix: "계약 api 수정"

검색기능 고려

* feat: "Querydsl을 활용한 주문 다중 검색 구현"

* feat: "검색어가 없어도 검색 가능"

* feat: "계약관리,주문상세 검색/페이징 구현"

* .gitignore fixed

* Fix: "거래처 상태 변경 수정"

* feat: set dockerfile

* Create homs-main.yml

* Update homs-main.yml

* Delete .github/workflows/homs-main.yml

* Create homs-main.yml

* Update homs-main.yml

* fix: gitignore and modify workflow

* Update homs-main.yml

* Update homs-main.yml

update java version

* Update homs-main.yml

* fix: modify git ignore

* Update homs-main.yml

* modify docker file in  homs-main.yml

* feat: add health check

* feat: add app spec file

* Fix: "거래처 등록 관련 수정"

거래처 등록 관련 ENUM 검증 부분 수정

* add deploy script homs-main.yml

* Update homs-main.yml

* fix: remove app spec file

* remove trigger pr homs-main.yml

* fix: add password update and transactional annotation

* Update README.md

프로그램 사양서, 단위 테스트 결과서, UI/UX 단위 테스트 업로드

* fix: "파일 다운 부분 수정"

* fix: "상품 파일 관련 불필요한 부분 리팩토링"

* Feat: "메뉴 및 상품 카테고리 CRUD 기능 수정"

메뉴 및 상품 카테고리 CRUD 기능 수정 (메뉴 buy, delivery, materials, sales 컬럼 추가)

* fix: "주문 상세 dto 변경"

* Fix: "유저dto수정"

UserResponseDto에서 기존 ID값으로 가져오는 방식에서 원활한 프론트 개발을 위해 Name을 직접 가져오는 방식으로 변경

* Fix: "거래처 상태 변경 수정 및 거래처 수정"

* feat: "이메일 세분화 및 템플릿 수정"

* feat: "이메일에 정산관리 추가"

* Fix: "사용자 삭제 여부"

* Fix: "계약요청dto수정"

id 요청에서 name그대로 주는걸로 수정해봤음

* fix: add set origin url

* feat: "클레임/주문 관리 구현"

* feat: "이메일 세분화 및 템플릿 수정"

* feat: "이메일에 정산관리 추가"

* feat: "클레임/주문 관리 구현"

* Update homs-main.yml

* Fix: "사용자 삭제 여부 기능 추가"

* feat: "클레임 상세 기능 구현중"

* fix: "클레임 상세 구현"

* Feat: "채팅v1"

백엔드 웹소켓 STOMP기반 채팅 기능 구현(테스트 필요)

* Fix: "채팅 엔드포인트 수정"

엔드포인트 및 컨피그 파일 수정

* Delete homs/src/main/resources/application-prod.yml 

해결!

* Fix: "채팅 JPA레포 수정"

메소드명 잘못되어있던 부분 수정

* feat: set req and res logging

* Feat: "정산 관리 POST API 추가"

* feat: set send error message

* Feat: "Delivery Address 추가"

Delivery Address 추가

* Fix: "채팅 v1 구현"

채팅 v1 구현

* Feat: "Delivery Address 완성 (User)"

Delivery Address 완성 (User)

* Feat: "배송 관리 address Name 추가"

* Feat: "배송 관리 Company Name 추가"

* Feat: "주문 관리 DeliveryInfo 추가"

* fix: set security exception handling

* feat: "배송지 상세 추가"

* Feat: "주문 관리 사용자별 DeliveryInfo 추가"

* fix: report log message

* feat: "이메일 템플릿 변경"

* feat: "이메일 템플릿 변경"

* fix: modify swagger description

* feat: "주문에 배송지 추가"
- 주문 엔티티와 주소 엔티티간의 연관관계를 느슨하게 설정

* Update README.md

시스템 아키텍처 최신화
CI/CD 계획서 링크 추가
통합 테스트 결과서 링크 추가

* Fix: "채팅 v2"

프론트에 맞게 채팅 리스트 가져올 수 있도록 수정

* Feat: "거래처 정보 회사 주소 정보 추가"

* Feat : "SideBar 메뉴 동적 생성 완료"

기존 코드는 SideBar_Back.vue 으로 옮겨둠

* fix: "정산 부분 이메일 구현 및 조금 수정"

* fix: "정산 부분 이메일 구현 및 조금 수정"

* Fix: "채팅관련 수정"

참여자 dto 생성 및 서비스 구현체 분리

* fix: user response dto add user role

* Update README.md

* fix: modify update user

* Fix: "채팅 v3"

채팅 프로토타입 완성, 채팅 진입점 이메일로 변경

* fix: "메뉴 불러오는 방식 변경"

* Feat: "주문건별 모든 상품 조회 api 생성"

부모 카테고리쪽에서 nullable에러가 나서 ProductCategoryMapper null에러처리

* fix: add log sender

* fix: "메일 이미지 첨부 문제 해결"

* fix: "계약 등록 수정"

* feat: "클레임 생성 수정중"

* fix: add set from eamil

* fix: modify mail health check

* fix: "하위주문 수정 작업"

* Feat: "클래임 상태 추가"

클래임에서 상태 값이 없어서 추가

* Update README.md

modify system architecture

* feat: "하위 주문 구현"

* feat: "클레임 생성 수정중"

* fix: "하위주문 수정 작업"

* feat: "하위 주문 구현"

* fix: remove jpa log and add err response req id

* fix: remove discord message send

---------

+++ homs/src/main/java/com/beyond/homs/product/service/ProductServiceImpl.java
@@ -2,6 +2,7 @@ package com.beyond.homs.product.service;
 
 import com.beyond.homs.common.exception.exceptions.CustomException;
 import com.beyond.homs.common.exception.messages.ExceptionMessage;
+import com.beyond.homs.common.util.SecurityUtil;
 import com.beyond.homs.product.dto.ProductFileRequestDto;
 import com.beyond.homs.product.dto.ProductFileResponseDto;
 import com.beyond.homs.product.dto.ProductListDto;
@@ -13,6 +14,7 @@ import com.beyond.homs.product.entity.ProductFile;
 import com.beyond.homs.product.repository.ProductCategoryRepository;
 import com.beyond.homs.product.repository.ProductFileRepository;
 import com.beyond.homs.product.repository.ProductRepository;
+import com.beyond.homs.user.data.UserRole;
 import com.beyond.homs.wms.repository.InventoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -44,6 +46,7 @@ public class ProductServiceImpl implements ProductService{
                 .productName(product.getProductName())
                 .productFeature(product.getProductFeature())
                 .productUsage(product.getProductUsage())
+                .productMinQuantity(product.getProductMinQuantity())
                 .category(product.getCategory())
                 .build();
     }
@@ -58,6 +61,12 @@ public class ProductServiceImpl implements ProductService{
     @Transactional
     @Override
     public ProductResponseDto createProduct(ProductRequestDto requestDto){
+
+        // 현재 유저의 권한을 가져옴
+        UserRole role = SecurityUtil.getCurrentUserRole();
+        if(role != UserRole.ROLE_ADMIN){
+            throw new CustomException(ExceptionMessage.NOT_PERMISSION_USER);
+        }
 
         ProductCategory productCategory = productCategoryRepository.findById(requestDto.getCategoryId())
                 .orElseThrow(() -> new CustomException(ExceptionMessage.CATEGORY_NOT_FOUND));
@@ -85,6 +94,13 @@ public class ProductServiceImpl implements ProductService{
     @Transactional
     @Override
     public ProductResponseDto updateProduct(Long productId, ProductRequestDto requestDto){
+
+        // 현재 유저의 권한을 가져옴
+        UserRole role = SecurityUtil.getCurrentUserRole();
+        if(role != UserRole.ROLE_ADMIN){
+            throw new CustomException(ExceptionMessage.NOT_PERMISSION_USER);
+        }
+
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new CustomException(ExceptionMessage.PRODUCT_NOT_FOUND));
 
@@ -109,6 +125,12 @@ public class ProductServiceImpl implements ProductService{
     @Override
     public void deleteProduct(Long productId){
 
+        // 현재 유저의 권한을 가져옴
+        UserRole role = SecurityUtil.getCurrentUserRole();
+        if(role != UserRole.ROLE_ADMIN){
+            throw new CustomException(ExceptionMessage.NOT_PERMISSION_USER);
+        }
+
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new CustomException(ExceptionMessage.PRODUCT_NOT_FOUND));
 
@@ -127,9 +149,6 @@ public class ProductServiceImpl implements ProductService{
                 .s3Image(productFile.getS3Image())
                 .s3Msds(productFile.getS3Msds())
                 .s3Tds1(productFile.getS3Tds1())
-                .s3Tds2(productFile.getS3Tds2())
-                .s3Property(productFile.getS3Property())
-                .s3Guide(productFile.getS3Guide())
                 .build();
     }
 
@@ -137,6 +156,12 @@ public class ProductServiceImpl implements ProductService{
     @Transactional
     @Override
     public void uploadProductFile(ProductFileRequestDto requestDto){
+
+        // 현재 유저의 권한을 가져옴
+        UserRole role = SecurityUtil.getCurrentUserRole();
+        if(role != UserRole.ROLE_ADMIN){
+            throw new CustomException(ExceptionMessage.NOT_PERMISSION_USER);
+        }
 
         Product product = productRepository.findById(requestDto.getProductId())
                 .orElseThrow(() -> new CustomException(ExceptionMessage.PRODUCT_NOT_FOUND));
@@ -146,9 +171,6 @@ public class ProductServiceImpl implements ProductService{
                 .s3Image(requestDto.getS3Image())
                 .s3Msds(requestDto.getS3Msds())
                 .s3Tds1(requestDto.getS3Tds1())
-                .s3Tds2(requestDto.getS3Tds2())
-                .s3Property(requestDto.getS3Property())
-                .s3Guide(requestDto.getS3Guide())
                 .build();
 
         productFileRepository.save(productFile);
@@ -158,6 +180,12 @@ public class ProductServiceImpl implements ProductService{
     @Transactional
     @Override
     public void updateProductFile(ProductFileRequestDto requestDto){
+        // 현재 유저의 권한을 가져옴
+        UserRole role = SecurityUtil.getCurrentUserRole();
+        if(role != UserRole.ROLE_ADMIN){
+            throw new CustomException(ExceptionMessage.NOT_PERMISSION_USER);
+        }
+
         ProductFile productFile = productFileRepository.findById(requestDto.getProductId())
                 .orElseThrow(() -> new CustomException(ExceptionMessage.PRODUCT_NOT_FOUND));
 
@@ -168,8 +196,13 @@ public class ProductServiceImpl implements ProductService{
     // 파일 삭제
     @Transactional
     @Override
-    public void deleteProductFile(Long productId)
-    {
+    public void deleteProductFile(Long productId) {
+        // 현재 유저의 권한을 가져옴
+        UserRole role = SecurityUtil.getCurrentUserRole();
+        if(role != UserRole.ROLE_ADMIN){
+            throw new CustomException(ExceptionMessage.NOT_PERMISSION_USER);
+        }
+
         productFileRepository.deleteById(productId);
     }
 }
